% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertTimeId.R
\name{convertTimeId}
\alias{convertTimeId}
\title{Convert an hourly timeId at a given time step}
\usage{
convertTimeId(hourId, timeStep, opts)
}
\arguments{
\item{hourId}{A numeric vector containing hourly time ids.}

\item{timeStep}{timeStep of the result.}

\item{opts}{Simulation options. It must contain elements \code{start} (first date
time of the study) and \code{firstWeekday}}
}
\value{
A numeric vector with same length as \code{hourId} reprensenting the
corresponding time ids at the required time step.
}
\description{
This function converts a vector of hourly time ids in time ids in another
time step.
}
\examples{
opts <- list(
  start = as.POSIXlt("2017-01-01", tz = "UTC"),
  firstWeekday = "Monday"
)

hourId <- 1:(24*14)
table(convertTimeId(hourId, "daily", opts))
table(convertTimeId(hourId, "weekly", opts))

}
