% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idCols.R
\name{getIdCols}
\alias{getIdCols}
\alias{addIdCol}
\title{Get and set Id columns}
\usage{
getIdCols(x, exclude = NULL)

addIdCol(colNames)
}
\arguments{
\item{x}{an AntaresDataTable.}

\item{exclude}{Columns to exclude in the result.}

\item{colNames}{Name of new ID columns to declare.}
}
\value{
\code{getIdCols()} returns a character vector containing the name of the ID
columns of an antaresDataTable. \code{addIdCol()} is only used for its side
effects and returns nothing.
}
\description{
\code{getIdCols()} returns the id columns of an AntaresDataTable. \code{addIdCol()}
permits to declare new ID columns. This can be useful when one wants to
define new ID columns in an \code{antaresDataTable} containing for instance
simulation name or day of week, etc.
}
