% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIniFile.R
\name{readIniFile}
\alias{readIniFile}
\alias{writeIniFile}
\title{Read and write an .ini file}
\usage{
readIniFile(file, stringsAsFactors = FALSE)

writeIniFile(x, file, overwrite = FALSE)
}
\arguments{
\item{file}{file path}

\item{stringsAsFactors}{Should string values treated as factors?}

\item{x}{Named list (see details).}

\item{overwrite}{Should the function overwrite the file if it already exists.}
}
\value{
A list with an element for each section of the .ini file.
}
\description{
These function read and write the content of an .ini file and converts to a list.
}
\details{
A .ini file has the following structure:\preformatted{[section1]
key1 = value1
key2 = value2
[section2]
key3 = value3
...
}

readIniFile converts it to a list with the following format:\preformatted{list(
  section1 = list(key1 = value1, key2 = value2),
  section2 = list(key3 = value3),
  ...
)
}

writeIniFile do the opposite operation.
}
