% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorderCols.R
\name{reorderCols}
\alias{reorderCols}
\title{reorder the columns of an antaresDataTable}
\usage{
reorderCols(x)
}
\arguments{
\item{x}{Object of class \code{antaresDataTable} or \code{data.table}.}
}
\value{
The function modifies by reference the order of the columns of its input. For
convenience it invisibly returns the modified input.
}
\description{
\code{reorderCols()} reorders the columns of an \code{antaresDataTable}. It puts first
the ID columns then the value columns.
}
\examples{
require(data.table)

mydata <- data.table(LOAD = rexp(5), timeId = 1:5, BALANCE = rnorm(5))
reorderCols(mydata)
names(mydata)

}
