% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeIdToDate.R
\name{timeIdToDate}
\alias{timeIdToDate}
\title{Convert a time id to a date}
\usage{
timeIdToDate(timeId, timeStep = c("hourly", "daily", "weekly", "monthly",
  "annual"), opts = simOptions())
}
\arguments{
\item{timeId}{A numeric vector}

\item{timeStep}{Character string indicating the time step of the parameter
\code{timeId}.}

\item{opts}{Simulation options. It must at least contain elements \code{start} and
and \code{firstWeekday}.}
}
\value{
A vector with same length as \code{timeId}. If \code{timeStep = "hourly"}, the result
is a \code{POSIXct} vector, else it is a \code{Date} vector.
}
\description{
This function can be used to convert times ids to dates. If time step is
hourly, it returns a POSIXct vector, else it always returns a date object.
For weekly, monthly and annual time step, it returns the first day of the
week, month or year.
}
